be/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
frozenlist==1.8.0
http_ece==1.2.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
multidict==6.7.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
pywebpush==2.1.1
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
zope.interface==8.0.1
